pyproject.toml
[project]
name = "ssh-key-mgr"
version = "0.1.2"
description = "Manager of SSH Keys"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Kalle M. Krog Aagaard", email = "git@k-moeller.dk"},
]
maintainers = [
    {name = "Kalle M. Krog Aagaard", email = "git@k-moeller.dk"},
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "ssh-proto-types>=0.1.3",
]

[project.urls]
Homepage = "https://github.com/KalleDK/py-ssh-key-mgr"
Documentation = "https://github.com/KalleDK/py-ssh-key-mgr"
Repository = "https://github.com/KalleDK/py-ssh-key-mgr.git"

[project.optional-dependencies]
cli = [
    "pyyaml>=6.0.3",
    "rich>=14.1.0",
    "typer>=0.19.2",
]
crypto = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "pycryptodome>=3.23.0",
]

[dependency-groups]
test = [
    "coverage>=7.10.7",
    "pydantic>=2.11.10",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
stubs = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssh_key_mgr"]

[tool.pyright]
include = ["src", "tests", "examples"]
typeCheckingMode = "strict"
exclude = ["**/__pycache__"]
pythonVersion = "3.13"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--cov=ssh_key_mgr",
    "--cov-report=html",
    "--cov-report=xml"
]
